fix: password confirm modals for admin add/remove
ask for password when toggling ownership of admin group
ask for password when kicking from admin group
allow passing headers to api.xxx methods
inviting user to admin group requires pwd reauth

+++ public/src/admin/manage/admins-mods.js
@@ -6,8 +6,13 @@ define('admin/manage/admins-mods', [
 	const AdminsMods = {};
 
 	AdminsMods.init = function () {
-		autocomplete.user($('#admin-search'), function (ev, ui) {
-			api.put(`/groups/administrators/membership/${ui.item.user.uid}`).then(() => {
+		autocomplete.user($('#admin-search'), async function (ev, ui) {
+			const password = await modals.promptPassword();
+			if (!password) {
+				return;
+			}
+
+			api.put(`/groups/administrators/membership/${ui.item.user.uid}`, { password }).then(() => {
 				$('#admin-search').val('');
 
 				if ($('.administrator-area [data-uid="' + ui.item.user.uid + '"]').length) {
@@ -20,19 +25,19 @@ define('admin/manage/admins-mods', [
 			}).catch(alerts.error);
 		});
 
-		$('.administrator-area').on('click', '.remove-user-icon', function () {
+		$('.administrator-area').on('click', '.remove-user-icon', async function () {
 			const userCard = $(this).parents('[data-uid]');
 			const uid = userCard.attr('data-uid');
 			if (parseInt(uid, 10) === parseInt(app.user.uid, 10)) {
 				return alerts.error('[[admin/manage/users:alerts.no-remove-yourself-admin]]');
 			}
-			modals.confirm('[[admin/manage/users:alerts.confirm-remove-admin]]', function (confirm) {
-				if (confirm) {
-					api.del(`/groups/administrators/membership/${uid}`).then(() => {
-						userCard.remove();
-					}).catch(alerts.error);
-				}
-			});
+			const password = await modals.promptPassword();
+			if (!password) {
+				return;
+			}
+			api.del(`/groups/administrators/membership/${uid}`, {}, { 'x-password-confirmation': password }).then(() => {
+				userCard.remove();
+			}).catch(alerts.error);
 		});
 
 		autocomplete.user($('#global-mod-search'), function (ev, ui) {
